added filtering for field-descriptor in factory
