+++ src/Sulu/Bundle/MediaBundle/Resources/translations/sulu/backend.de.xlf
@@ -30,6 +30,10 @@
                 <source>media.media.size</source>
                 <target>Dateigröße</target>
             </trans-unit>
+            <trans-unit id="3c8f1c7c02abd468cc0794e987e2b2ee" resname="media.media.mime-type">
+                <source>media.media.mime-type</source>
+                <target>MIME-Type</target>
+            </trans-unit>
             <trans-unit id="6f5e8f5b65fd4e7ae7ec509e742cc85f" resname="media.media.thumbnails">
                 <source>media.media.thumbnails</source>
                 <target>Vorschaubild</target>
