+++ src/Sulu/Component/Rest/RestHelperInterface.php
@@ -11,7 +11,7 @@
 
 namespace Sulu\Component\Rest;
 
-use Sulu\Component\Rest\ListBuilder\AbstractFieldDescriptor;
+use Sulu\Component\Rest\ListBuilder\FieldDescriptor;
 use Sulu\Component\Rest\ListBuilder\ListBuilderInterface;
 use Traversable;
 
@@ -21,7 +21,7 @@ interface RestHelperInterface
      * Initializes the given ListBuilder with the standard values from the request.
      *
      * @param ListBuilderInterface      $listBuilder      The ListBuilder to initialize
-     * @param AbstractFieldDescriptor[] $fieldDescriptors The FieldDescriptors available for this object type
+     * @param FieldDescriptor[] $fieldDescriptors The FieldDescriptors available for this object type
      */
     public function initializeListBuilder(ListBuilderInterface $listBuilder, array $fieldDescriptors);
