+++ src/Sulu/Component/Rest/Tests/Functional/ListBuilder/Metadata/FieldDescriptorFactoryTest.php
@@ -20,6 +20,7 @@ use Sulu\Component\Rest\ListBuilder\Doctrine\FieldDescriptor\DoctrineFieldDescri
 use Sulu\Component\Rest\ListBuilder\Doctrine\FieldDescriptor\DoctrineGroupConcatFieldDescriptor;
 use Sulu\Component\Rest\ListBuilder\Doctrine\FieldDescriptor\DoctrineIdentityFieldDescriptor;
 use Sulu\Component\Rest\ListBuilder\Doctrine\FieldDescriptor\DoctrineJoinDescriptor;
+use Sulu\Component\Rest\ListBuilder\FieldDescriptor;
 use Sulu\Component\Rest\ListBuilder\FieldDescriptorInterface;
 use Sulu\Component\Rest\ListBuilder\Metadata\Doctrine\Driver\XmlDriver as DoctrineXmlDriver;
 use Sulu\Component\Rest\ListBuilder\Metadata\FieldDescriptorFactory;
@@ -284,6 +285,42 @@ class FieldDescriptorFactoryTest extends \PHPUnit_Framework_TestCase
         $fieldDescriptor = $factory->getFieldDescriptorForClass(\stdClass::class);
 
         $this->assertEmpty($fieldDescriptor);
+    }
+
+    public function testGetFieldDescriptorForClassMixed()
+    {
+        $this->locator->findFileForClass(new \ReflectionClass(new \stdClass()), 'xml')
+            ->willReturn(__DIR__ . '/Resources/mixed.xml');
+
+        $provider = new ChainProvider($this->chain);
+        $factory = new FieldDescriptorFactory($provider, $this->configCachePath, $this->debug);
+        $fieldDescriptor = $factory->getFieldDescriptorForClass(\stdClass::class);
+
+        $this->assertCount(2, $fieldDescriptor);
+        $this->assertFieldDescriptor(
+            ['name' => 'id', 'translation' => 'Id', 'disabled' => true],
+            $fieldDescriptor['id']
+        );
+        $this->assertFieldDescriptor(
+            ['name' => 'name', 'translation' => 'Name', 'disabled' => true, 'instance' => FieldDescriptor::class],
+            $fieldDescriptor['name']
+        );
+    }
+
+    public function testGetFieldDescriptorForClassMixedByType()
+    {
+        $this->locator->findFileForClass(new \ReflectionClass(new \stdClass()), 'xml')
+            ->willReturn(__DIR__ . '/Resources/mixed.xml');
+
+        $provider = new ChainProvider($this->chain);
+        $factory = new FieldDescriptorFactory($provider, $this->configCachePath, $this->debug);
+        $fieldDescriptor = $factory->getFieldDescriptorForClass(\stdClass::class, [], DoctrineFieldDescriptor::class);
+
+        $this->assertCount(1, $fieldDescriptor);
+        $this->assertFieldDescriptor(
+            ['name' => 'id', 'translation' => 'Id', 'disabled' => true],
+            $fieldDescriptor['id']
+        );
     }
 
     private function assertFieldDescriptors(array $expected, array $fieldDescriptors)
