+++ src/Sulu/Component/Rest/Tests/Functional/ListBuilder/Metadata/Resources/mixed.xml
@@ -1,0 +1,11 @@
+<class xmlns="http://schemas.sulu.io/class/general"
+       xmlns:orm="http://schemas.sulu.io/class/doctrine">
+    <properties>
+        <property name="id">
+            <orm:field-name>id</orm:field-name>
+            <orm:entity-name>%sulu.model.contact.class%</orm:entity-name>
+        </property>
+
+        <property name="name"/>
+    </properties>
+</class>
